requirements.txt
streamlit
psycopg2-binary
joblib
xgboost
plotly
langchain
ollama
